# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Football Pool API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/picks": {
      "get": {
        "summary": "Get user picks",
        "operationId": "getPicks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/picks/submit": {
      "post": {
        "summary": "Submit picks",
        "operationId": "submitPicks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PickRequest"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PickResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/picks/submit": {
      "post": {
        "summary": "Admin submit picks",
        "operationId": "adminSubmitPicks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PickRequest"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PickResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/survivor/picks": {
      "get": {
        "summary": "Get survivor picks",
        "operationId": "getSurvivorPicks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SurvivorPickResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/picks": {
      "get": {
        "summary": "Admin list picks",
        "operationId": "adminListPicks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/picks/week/{week}": {
      "get": {
        "summary": "Admin get picks by week",
        "operationId": "adminGetPicksByWeek",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "week",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/picks/user/{userId}": {
      "get": {
        "summary": "Admin get picks by user",
        "operationId": "adminGetPicksByUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/picks/{id}": {
      "delete": {
        "summary": "Admin delete pick",
        "operationId": "adminDeletePick",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserResponse": {
        "type": "object",
        "required": ["id", "name", "email", "role", "created_at", "updated_at"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerResponse"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PickListResponse": {
        "type": "object",
        "required": ["picks", "pagination"],
        "properties": {
          "picks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PickResponse"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponse"
          }
        }
      },
...
      "PlayerResponse": {
        "type": "object",
        "required": ["id", "user_id", "name", "address"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint"
          },
          "user_id": {
            "type": "integer",
            "format": "uint"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
...
      "GameResponse": {
        "type": "object",
        "required": ["id", "week", "season", "favorite_team", "underdog_team", "spread", "created_at", "updated_at"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint"
          },
          "week": {
            "type": "integer"
          },
          "season": {
            "type": "integer"
          },
          "favorite_team": {
            "type": "string"
          },
          "underdog_team": {
            "type": "string"
          },
          "spread": {
            "type": "number",
            "format": "float"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PickResponse": {
        "type": "object",
        "required": ["id", "user_id", "game_id", "picked", "rank", "quick_pick", "created_at", "updated_at"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint"
          },
          "user_id": {
            "type": "integer",
            "format": "uint"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "game_id": {
            "type": "integer",
            "format": "uint"
          },
          "game": {
            "$ref": "#/components/schemas/GameResponse"
          },
          "picked": {
            "type": "string"
          },
          "rank": {
            "type": "integer"
          },
          "quick_pick": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PickRequest": {
        "type": "object",
        "required": ["game_id", "picked", "rank", "quick_pick"],
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "uint"
          },
          "game_id": {
            "type": "integer",
            "format": "uint"
          },
          "picked": {
            "type": "string"
          },
          "rank": {
            "type": "integer"
          },
          "quick_pick": {
            "type": "boolean"
          }
        }
      },
...
      "SurvivorPickResponse": {
        "type": "object",
        "required": ["id", "user_id", "week", "team", "created_at", "updated_at"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint"
          },
          "user_id": {
            "type": "integer",
            "format": "uint"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "week": {
            "type": "integer"
          },
          "team": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginationResponse": {
        "type": "object",
        "required": ["page", "limit", "total", "pages"],
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}